Format Solidity + JS + TS files, pnpm workspace

+++ contracts/src/Interfaces/ICollSurplusPool.sol
@@ -2,19 +2,15 @@
 
 pragma solidity 0.8.18;
 
-
 interface ICollSurplusPool {
-    function setAddresses(
-        address _borrowerOperationsAddress,
-        address _troveManagerAddress,
-        address _activePoolAddress
-    ) external;
+    function setAddresses(address _borrowerOperationsAddress, address _troveManagerAddress, address _activePoolAddress)
+        external;
 
-    function getETHBalance() external view returns (uint);
+    function getETHBalance() external view returns (uint256);
 
-    function getCollateral(uint256 _troveId) external view returns (uint);
+    function getCollateral(uint256 _troveId) external view returns (uint256);
 
-    function accountSurplus(uint256 _troveId, uint _amount) external;
+    function accountSurplus(uint256 _troveId, uint256 _amount) external;
 
     function claimColl(address _account, uint256 _troveId) external;
 }
