+++ contracts/src/Interfaces/ICommunityIssuance.sol
@@ -2,10 +2,10 @@
 
 pragma solidity 0.8.18;
 
-interface ICommunityIssuance { 
+interface ICommunityIssuance {
     function setAddresses(address _lqtyTokenAddress, address _stabilityPoolAddress) external;
 
-    function issueLQTY() external returns (uint);
+    function issueLQTY() external returns (uint256);
 
-    function sendLQTY(address _account, uint _LQTYamount) external;
+    function sendLQTY(address _account, uint256 _LQTYamount) external;
 }
